@@ -7,8 +7,8 @@ import { reportWebVitals } from '@/reportWebVitals';
 
 const container = document.getElementById('root');
 
-// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+ 
 const root = createRoot(container!);
 root.render(<App />);
 
-reportWebVitals();
+reportWebVitals(console.log);
